failover: delay error handling when no servers are configured
The BitBoxApp (primary client of this library) currently has no good
way of handling and surfacing errors when `failover.New()` fails.

In case no servers are configured, instead of failing there, we treat
it as a connectoin error, so the BitBoxApp can display it as a normal
connection error.

This also makes it easier to add functionality in the future to
dynamically change the servers, i.e. add servers if there are no
servers.