failover: fix unit test race condition
The OnDisconnect callback set up in one test could be called after
that unit test has finished and a new one has started, wrongly
modifying the `onDisconnectCount` counter of the new test.

Similar issue: since OnDisconnect is called in a goroutine, the order
in which they are executed is not deterministic.