failover: add method to manually reconnect
When all servers are down, a retry only happens after the retry
timeout. We want to be able to manually trigger reconnecting earlier
than that.

When the BitBoxApp iOS app goes into background, it kills TCP
connections shortly after. We want to be able to trigger a reconnect
as soon as the iOS app comes to the foreground again.