failover: close immediately
A failover.Close() call would not abort all operations immediately -
any call that is in the middle of failing over would wait for the
remaining retryTimeout and try to connect again before either
sucdeeding and finishing the call, or finally returning `ErrClosed`.

This can cause a long delay in shutting down the client.

With this commit, we abort with `ErrClosed` immediately in this case.